added styles using Reactstrap
