+++ package.json
@@ -6,9 +6,11 @@
     "@testing-library/jest-dom": "^5.11.5",
     "@testing-library/react": "^11.1.0",
     "@testing-library/user-event": "^12.1.10",
+    "bootstrap": "^4.5.3",
     "react": "^17.0.1",
     "react-dom": "^17.0.1",
     "react-scripts": "4.0.0",
+    "reactstrap": "^8.7.0",
     "web-vitals": "^0.2.4"
   },
   "scripts": {
